requirements.txt
# Unified requirements for all course exercises (Lessons 1-5)

# Core dependency - Google Generative AI SDK (Vertex AI / Gemini)
google-genai>=1.62.0

# Data handling and validation
dataclasses-json>=0.5.9
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0




